migrate test_bigip to pytest
